# perf: metainfo reduce mallocgc