--- a/cloud/metainfo/utils.go
+++ b/cloud/metainfo/utils.go
@@ -28,48 +28,81 @@ func HasMetaInfo(ctx context.Context) bool {
 // Only those keys with prefixes defined in this module would be used.
 // If the context has been carrying metanifo pairs, they will be merged as a basis.
 func SetMetaInfoFromMap(ctx context.Context, m map[string]string) context.Context {
-	if ctx == nil {
-		return nil
-	}
-
-	if len(m) == 0 {
+	if ctx == nil || len(m) == 0 {
 		return ctx
 	}
+	var persistent kvstore
+	var transient kvstore
+	var stale kvstore
 
-	var mv *mapView
-	if x := getNode(ctx); x != nil {
-		mv = x.mapView()
+	// inherit from exist ctx node
+	mapSize := len(m)
+	nd := getNode(ctx)
+	inherit := nd != nil
+	if inherit {
+		// inherit from node
+		persistent = newKVStore(mapSize)
+		transient = newKVStore(mapSize)
+		stale = newKVStore(mapSize)
+		sliceToMap(nd.persistent, persistent)
+		sliceToMap(nd.transient, transient)
+		sliceToMap(nd.stale, stale)
 	} else {
-		mv = newMapView()
+		// make new node
+		nd = &node{
+			persistent: make([]kv, 0, mapSize),
+			transient:  make([]kv, 0, mapSize),
+			stale:      make([]kv, 0, mapSize),
+		}
 	}
 
+	// insert new kvs from m to node
 	for k, v := range m {
 		if len(k) == 0 || len(v) == 0 {
 			continue
 		}
-
 		switch {
 		case strings.HasPrefix(k, PrefixTransientUpstream):
 			if len(k) > lenPTU { // do not move this condition to the case statement to prevent a PTU matches PT
-				mv.stale[k[lenPTU:]] = v
+				if inherit {
+					stale[k[lenPTU:]] = v
+				} else {
+					nd.stale = append(nd.stale, kv{key: k[lenPTU:], val: v})
+				}
 			}
 		case strings.HasPrefix(k, PrefixTransient):
 			if len(k) > lenPT {
-				mv.transient[k[lenPT:]] = v
+				if inherit {
+					transient[k[lenPT:]] = v
+				} else {
+					nd.transient = append(nd.transient, kv{key: k[lenPT:], val: v})
+				}
 			}
 		case strings.HasPrefix(k, PrefixPersistent):
 			if len(k) > lenPP {
-				mv.persistent[k[lenPP:]] = v
+				if inherit {
+					persistent[k[lenPP:]] = v
+				} else {
+					nd.persistent = append(nd.persistent, kv{key: k[lenPP:], val: v})
+				}
 			}
 		}
 	}
 
-	if mv.size() == 0 {
-		// TODO: remove this?
+	if nd.size() == 0 { // return original ctx if no invalid key in map
 		return ctx
 	}
-
-	return withNode(ctx, mv.toNode())
+	// transfer map to list
+	if inherit {
+		nd.stale = stale.toList(nd.stale)
+		nd.transient = transient.toList(nd.transient)
+		nd.persistent = persistent.toList(nd.persistent)
+		stale.recycle()
+		transient.recycle()
+		persistent.recycle()
+	}
+	ctx = withNode(ctx, nd)
+	return ctx
 }
 
 // SaveMetaInfoToMap set key-value pairs from ctx to m while filtering out transient-upstream data.
@@ -78,29 +111,28 @@ func SaveMetaInfoToMap(ctx context.Context, m map[string]string) {
 		return
 	}
 	ctx = TransferForward(ctx)
-	for k, v := range GetAllValues(ctx) {
-		m[PrefixTransient+k] = v
-	}
-	for k, v := range GetAllPersistentValues(ctx) {
-		m[PrefixPersistent+k] = v
+	if n := getNode(ctx); n != nil {
+		for _, kv := range n.persistent {
+			m[PrefixPersistent+kv.key] = kv.val
+		}
+		for _, kv := range n.transient {
+			m[PrefixTransient+kv.key] = kv.val
+		}
+		for _, kv := range n.stale {
+			m[PrefixTransient+kv.key] = kv.val
+		}
 	}
 }
 
-// sliceToMap converts a kv slice to map. If the slice is empty, an empty map will be returned instead of nil.
-func sliceToMap(slice []kv) (m map[string]string) {
-	if size := len(slice); size == 0 {
-		m = make(map[string]string)
-	} else {
-		m = make(map[string]string, size)
-	}
+// sliceToMap converts a kv slice to map.
+func sliceToMap(slice []kv, kvs kvstore) {
 	for _, kv := range slice {
-		m[kv.key] = kv.val
+		kvs[kv.key] = kv.val
 	}
-	return
 }
 
 // mapToSlice converts a map to a kv slice. If the map is empty, the return value will be nil.
-func mapToSlice(kvs map[string]string) (slice []kv) {
+func mapToSlice(kvs kvstore) (slice []kv) {
 	size := len(kvs)
 	if size == 0 {
 		return