--- a/cloud/metainfo/utils.go
+++ b/cloud/metainfo/utils.go
@@ -112,14 +112,14 @@ func SaveMetaInfoToMap(ctx context.Context, m map[string]string) {
 	}
 	ctx = TransferForward(ctx)
 	if n := getNode(ctx); n != nil {
-		for _, kv := range n.persistent {
-			m[PrefixPersistent+kv.key] = kv.val
+		for _, kv := range n.stale {
+			m[PrefixTransient+kv.key] = kv.val
 		}
 		for _, kv := range n.transient {
 			m[PrefixTransient+kv.key] = kv.val
 		}
-		for _, kv := range n.stale {
-			m[PrefixTransient+kv.key] = kv.val
+		for _, kv := range n.persistent {
+			m[PrefixPersistent+kv.key] = kv.val
 		}
 	}
 }