### Summary

- FromHTTPHeader: reduce 18%~46%
- SaveMetaInfoToMap: reduced 37%~40%
- SetMetaInfoFromMap: reduced 60%~66%

### Benchmark

Before:
```
BenchmarkFromHTTPHeader/FromHTTPHeader-10-4         	  533098	      2219 ns/op	    1177 B/op	      19 allocs/op
BenchmarkFromHTTPHeader/FromHTTPHeader-20-4         	  239277	      5008 ns/op	    2683 B/op	      31 allocs/op
BenchmarkAll/SaveMetaInfoToMap_10-4                 	  235621	      4867 ns/op	    3880 B/op	      28 allocs/op
BenchmarkAll/SaveMetaInfoToMap_20-4                 	  121765	     10373 ns/op	    8345 B/op	      50 allocs/op
BenchmarkAll/SaveMetaInfoToMap_50-4                 	   50384	     23848 ns/op	   18977 B/op	     118 allocs/op
BenchmarkAll/SaveMetaInfoToMap_100-4                	   23650	     50333 ns/op	   38275 B/op	     226 allocs/op
BenchmarkAll/SetMetaInfoFromMap_10-4                	  377126	      3019 ns/op	    1932 B/op	       6 allocs/op
BenchmarkAll/SetMetaInfoFromMap_20-4                	  178387	      6894 ns/op	    5111 B/op	       8 allocs/op
BenchmarkAll/SetMetaInfoFromMap_50-4                	   76599	     16080 ns/op	   13111 B/op	      14 allocs/op
BenchmarkAll/SetMetaInfoFromMap_100-4               	   38401	     31900 ns/op	   26885 B/op	      22 allocs/op
```

After: 
```
BenchmarkFromHTTPHeader/FromHTTPHeader-10-4         	  723223	      1659 ns/op	    1737 B/op	      14 allocs/op
BenchmarkFromHTTPHeader/FromHTTPHeader-20-4         	  440990	      2764 ns/op	    1769 B/op	      24 allocs/op
BenchmarkAll/SaveMetaInfoToMap_10-4                 	  369705	      3007 ns/op	    2619 B/op	      24 allocs/op
BenchmarkAll/SaveMetaInfoToMap_20-4                 	  188844	      6331 ns/op	    5851 B/op	      46 allocs/op
BenchmarkAll/SaveMetaInfoToMap_50-4                 	   80311	     14951 ns/op	   13504 B/op	     111 allocs/op
BenchmarkAll/SaveMetaInfoToMap_100-4                	   39208	     30654 ns/op	   27362 B/op	     216 allocs/op
BenchmarkAll/SetMetaInfoFromMap_10-4                	  934087	      1296 ns/op	    2048 B/op	       5 allocs/op
BenchmarkAll/SetMetaInfoFromMap_20-4                	  529394	      2293 ns/op	    3968 B/op	       5 allocs/op
BenchmarkAll/SetMetaInfoFromMap_50-4                	  237283	      5076 ns/op	    9728 B/op	       5 allocs/op
BenchmarkAll/SetMetaInfoFromMap_100-4               	  128302	      9549 ns/op	   19712 B/op	       5 allocs/op
```